Add initial WZX support for Pokémon XD Gale of Darkness
Mainly "traced" what the loading code in GXXP01 does.
Seems kind of weird format with those 0x70 "chunks", maybe the header is
also one (since the size it skips matches).
Maybe look later at the actual usage if that is the case and maybe
simplify, if needed.

Also needed to "patch" GSscene because in one file
(wzx_carde_bg03.fsys/0CE72000_carde_bg03.wzx starting at 0xA0)
the JObj offset is 0x00000000 (which is valid for some other files),
but not in that file. So a "workaround" to try to detect that file has
been implemented. It appears to only contain motion data, so maybe add a
new file type and modify GSscene that when processing motion scenes to
ignore JObj models. But now now. Later. Maybe.